data-aquisition/requirements.txt
requests
redis
python-dotenv
google-genai
beautifulsoup4
lxml
colorama